move workspace header status spinner to own component
